+++ scripts/dev.sh
@@ -507,6 +507,7 @@ elif [ "$MODE" = "test" ]; then
     export HASURA_GRAPHQL_PG_SOURCE_URL_1=${HASURA_GRAPHQL_PG_SOURCE_URL_1-$PG_DB_URL}
     export HASURA_GRAPHQL_PG_SOURCE_URL_2=${HASURA_GRAPHQL_PG_SOURCE_URL_2-$PG_DB_URL}
     export HASURA_GRAPHQL_EXPERIMENTAL_FEATURES="inherited_roles"
+    export HASURA_GRAPHQL_MSSQL_SOURCE_URL=$MSSQL_CONN_STR
 
     # Using --metadata-database-url flag to test multiple backends
     #       HASURA_GRAPHQL_PG_SOURCE_URL_* For a couple multi-source pytests: